feat: create hook authentication
