+++ src/App.tsx
@@ -1,67 +1,12 @@
-import React, { createContext, useState, useEffect } from 'react';
 import Routers from "./components/Routers";
-import { auth, firebase } from './services/firebase';
 
-
-type User = {
-  id: string;
-  name: string;
-  avatar: string;
-}
-
-type AuthContextType = {
-  user: User | undefined;
-  signInWithGoogle: () => Promise<void>;
-}
-
-export const AuthContext = createContext({} as AuthContextType);
+import { AuthContextProvider } from './contexts/AuthContext';
 
 function App() {
-  const [user, setUser] = useState<User>();
-
-  useEffect(() => {
-    auth.onAuthStateChanged(user => {
-      if (user) {
-        const { displayName, photoURL, uid } = user;
-      
-        if(!displayName || !photoURL) {
-          throw new Error("Missing information from google Account.");
-        }
-
-        setUser({
-          id: uid,
-          name: displayName,
-          avatar: photoURL,
-        });
-      }
-    });
-  }, []);
-
-  async function signInWithGoogle() {
-    const provider = new firebase.auth.GoogleAuthProvider();
-
-    const result = await auth.signInWithPopup(provider);
-
-    if(result.user) {
-      const { displayName, photoURL, uid } = result.user;
-      
-      if(!displayName || !photoURL) {
-        throw new Error("Missing information from google Account.");
-    }
-
-    setUser({
-      id: uid,
-      name: displayName,
-      avatar: photoURL,
-    });
-
-    }
-  }
-
   return (
-    <AuthContext.Provider value={{ user, signInWithGoogle }}>
+    <AuthContextProvider>
       <Routers />
-    </AuthContext.Provider>
+    </AuthContextProvider>
   );
 }
