+++ src/pages/NewRoom.tsx
@@ -4,11 +4,11 @@ import illustrationImg from '../assets/images/illustration.svg';
 import logoImg from '../assets/images/logo.svg';
 import '../styles/auth.scss';
 import { Button } from '../components/Button';
-import { useContext } from 'react';
-import { AuthContext } from '../App';
+// import { useContext } from 'react';
+// import { AuthContext } from '../contexts/AuthContext';
 
 export function NewRoom() {
-  const { user } = useContext(AuthContext);
+  // const { user } = useContext(AuthContext);
 
   return (
     <div id="page-auth">
@@ -21,7 +21,6 @@ export function NewRoom() {
       <main>
         <div className="main-content">
           <img src={logoImg} alt="Letmeask"/>
-          <h1>{user?.name}</h1>
           <h2>Criar uma nova sala</h2>
          <form>
             <input
